removing footer
